Update delete file

+++ web/public/javascripts/components.js
@@ -8838,18 +8838,17 @@
 
     /**
      * 生成文件夹的行界面。
-     * @param {Directory} folder 
-     * @param {boolean} extended 
+     * @param {Directory|TrashDirectory} folder 
      */
-    function makeFolderRow(folder, extended) {
+    function makeFolderRow(folder) {
         var id = folder.getId();
         var name = folder.getName();
         var time = folder.getLastModified();
-        if (extended) {
+        if (folder instanceof TrashDirectory) {
             name = name + ' （于 ' + g.formatYMDHMS(folder.getTrashTimestamp()) + '）';
         }
 
-        return [
+        var html = [
             '<tr ondblclick="app.filePanel.changeDirectory(\'', id, '\')" id="ftr_', id, '">',
                 '<td onclick="app.filePanel.toggleSelect(\'', id, '\')">', '<div class="icheck-primary">',
                     '<input type="checkbox" data-type="folder" id="', id, '">',
@@ -8859,29 +8858,38 @@
                 '<td class="file-size">--</td>',
                 '<td class="file-lastmodifed">', g.formatYMDHMS(time), '</td>',
                 '<td class="file-completion">--</td>',
-                '<td class="file-operate">',
-                    '<button onclick="javascript:app.filePanel.renameDirectory(', id, ');"',
-                        ' type="button" class="btn btn-primary btn-sm" title="重命名"><i class="fas fa-edit"></i></button>',
-                    '<button onclick="javascript:app.filePanel.promptDeleteDirectory(', id, ');"',
-                        ' type="button" class="btn btn-danger btn-sm" title="删除"><i class="far fa-trash-alt"></i></button>',
-                '</td>',
-            '</tr>'
+                '<td class="file-operate">'
         ];
+
+        if (!(folder instanceof TrashDirectory)) {
+            html = html.concat([
+                '<button onclick="javascript:app.filePanel.renameDirectory(', id, ');"',
+                    ' type="button" class="btn btn-primary btn-sm" title="重命名"><i class="fas fa-edit"></i></button>'
+            ]);
+        }
+
+        html = html.concat([
+            '<button onclick="javascript:app.filePanel.promptDeleteDirectory(', id, ');"',
+                ' type="button" class="btn btn-danger btn-sm" title="删除"><i class="far fa-trash-alt"></i></button>'
+        ]);
+
+        html.push('</td></tr>');
+
+        return html;
     }
 
     /**
      * 生成文件的行界面。
      * @param {*} fileLabel 
-     * @param {*} extended 
      */
-    function makeFileRow(fileLabel, extended) {
+    function makeFileRow(fileLabel) {
+        var id = fileLabel.getId();
         var name = fileLabel.getFileName();
-        if (extended) {
+        if (fileLabel instanceof TrashFile) {
             name = name + ' （于 ' + g.formatYMDHMS(fileLabel.getTrashTimestamp()) + '）';
         }
 
-        var id = fileLabel.getId();
-        return [
+        var html = [
             '<tr ondblclick="app.filePanel.openFileDetails(\'', fileLabel.getFileCode(), '\')" id="ftr_', id, '">',
                 '<td onclick="app.filePanel.toggleSelect(\'', id, '\')">', '<div class="icheck-primary">',
                     '<input type="checkbox" data-type="file" id="', id, '">',
@@ -8891,25 +8899,38 @@
                 '<td class="file-size">', g.formatSize(fileLabel.getFileSize()), '</td>',
                 '<td class="file-lastmodifed">', g.formatYMDHMS(fileLabel.getLastModified()), '</td>',
                 '<td class="file-completion">', g.formatYMDHMS(fileLabel.getCompletionTime()), '</td>',
-                '<td class="file-operate">',
-                    '<button ', 'onclick="app.filePanel.downloadFile(\'', fileLabel.getFileCode(), '\')"',
+                '<td class="file-operate">'
+        ];
+
+        if (fileLabel instanceof TrashFile) {
+            html = html.concat([
+                '<button ', 'onclick="app.filePanel.promptDeleteFile(\'', fileLabel.getFileName(), '\',\'', fileLabel.getFileCode(), '\')"',
+                    ' type="button" class="btn btn-danger btn-sm" title="删除文件" data-target="recycle-file"><i class="far fa-trash-alt"></i></button>'
+            ]);
+        }
+        else {
+            html = html.concat([
+                '<button ', 'onclick="app.filePanel.downloadFile(\'', fileLabel.getFileCode(), '\')"',
                         ' type="button" class="btn btn-primary btn-sm" title="下载"><i class="fas fa-download"></i></button>',
                     '<button ', 'onclick="app.filePanel.openCreateSharingTagDialog(\'', fileLabel.getFileCode(), '\')"',
                         ' type="button" class="btn btn-info btn-sm" title="分享" data-target="share-file"><i class="fas fa-share-alt"></i></button>',
-                    '<div class="btn-group">',
-                        '<button type="button" class="btn btn-secondary btn-sm dropdown-toggle dropdown-icon" data-toggle="dropdown" title="更多操作">','</button>',
-                        '<div class="dropdown-menu">',
-                            '<a class="dropdown-item text-sm" href="javascript:app.filePanel.renameFile(\'', fileLabel.getFileName(), '\',\'', fileLabel.getFileCode(), '\');">',
-                                '<i class="far fa-edit"></i>&nbsp;&nbsp;重命名</a>',
-                            '<a class="dropdown-item text-sm" href="javascript:app.filePanel.openFolderDialog(\'', fileLabel.getFileName(), '\',\'', fileLabel.getFileCode(), '\');">',
-                                '<i class="far fa-folder"></i>&nbsp;&nbsp;移动文件</a>',
-                            '<a class="dropdown-item text-sm" href="javascript:app.filePanel.promptDeleteFile(\'', fileLabel.getFileName(), '\',\'', fileLabel.getFileCode(), '\');">', 
-                                '<span class="text-danger"><i class="far fa-trash-alt"></i>&nbsp;&nbsp;删除文件<span></a>',
-                        '</div>',
+                '<div class="btn-group">',
+                    '<button type="button" class="btn btn-secondary btn-sm dropdown-toggle dropdown-icon" data-toggle="dropdown" title="更多操作">','</button>',
+                    '<div class="dropdown-menu">',
+                        '<a class="dropdown-item text-sm" href="javascript:app.filePanel.renameFile(\'', fileLabel.getFileName(), '\',\'', fileLabel.getFileCode(), '\');">',
+                            '<i class="far fa-edit"></i>&nbsp;&nbsp;重命名</a>',
+                        '<a class="dropdown-item text-sm" href="javascript:app.filePanel.openFolderDialog(\'', fileLabel.getFileName(), '\',\'', fileLabel.getFileCode(), '\');">',
+                            '<i class="far fa-folder"></i>&nbsp;&nbsp;移动文件</a>',
+                        '<a class="dropdown-item text-sm" href="javascript:app.filePanel.promptDeleteFile(\'', fileLabel.getFileName(), '\',\'', fileLabel.getFileCode(), '\');">', 
+                            '<span class="text-danger"><i class="far fa-trash-alt"></i>&nbsp;&nbsp;删除文件<span></a>',
                     '</div>',
-                '</td>',
-            '</tr>'
-        ];
+                '</div>'
+            ]);
+        }
+
+        html.push('</td></tr>');
+
+        return html;
     }
 
     /**
@@ -9068,9 +9089,8 @@
     /**
      * 更新表格数据。
      * @param {Array} list 数据列表。
-     * @param {boolean} [extended] 是否在文件名后附加目录信息。
      */
-    FileTable.prototype.updatePage = function(list, extended) {
+    FileTable.prototype.updatePage = function(list) {
         if (list.length == 0) {
             surface[0].innerHTML = '';
             noFileBg.css('display', 'block');
@@ -9082,13 +9102,13 @@
         list.forEach(function(element) {
             var rowHtml = null;
             if (element instanceof FileLabel || element instanceof TrashFile) {
-                rowHtml = makeFileRow(element, extended);
+                rowHtml = makeFileRow(element);
             }
             else if (element instanceof SearchItem) {
-                rowHtml = makeSearchItemRow(element, extended);
+                rowHtml = makeSearchItemRow(element);
             }
             else {
-                rowHtml = makeFolderRow(element, extended);
+                rowHtml = makeFolderRow(element);
             }
 
             html = html.concat(rowHtml);
@@ -9809,7 +9829,7 @@
     /**
      * 使用当前目录数据刷新表格。
      * @param {boolean} reset 是否重置当前目录。
-     * @param {function} [completion] 完成。
+     * @param {function} [completion] 完成操作回调。
      */
     FilePanel.prototype.refreshTable = function(reset, completion) {
         if (reset) {
@@ -10038,8 +10058,8 @@
         btnPrev.attr('disabled', 'disabled');
         btnNext.attr('disabled', 'disabled');
 
-        window.cube().fs.listTrash(0, 20, function(root, list, begin, end) {
-            table.updatePage(list, true);
+        window.cube().fs.listTrash(0, 19, function(root, list, begin, end) {
+            table.updatePage(list);
 
             refreshPageInfoBar(list.length);
 
@@ -10126,6 +10146,9 @@
         var directory = null;
 
         if (selectedRecycleBin) {
+            // 回收站里的文件
+            //fileLabel = g.cube().fs.queryTrashFile(fileCode);
+            return;
         }
         else if (selectedFilter) {
             var searchItem = window.cube().fs.querySearch(directoryId, fileCode);
@@ -10282,26 +10305,56 @@
      * @param {number} dirId 
      */
     FilePanel.prototype.promptDeleteDirectory = function(dirId) {
-        var dir = g.cube().fs.queryDirectory(dirId);
-        if (null == dir) {
-            alert('查找目录出错');
-            return;
-        }
-
-        var text = ['您确定要删除文件夹 ', '“<span class="text-danger">', dir.getName(), '</span>” 及该文件夹内的',
-                    '<span class="text-danger">全部文件</span>',
-                    '吗？'];
-        g.dialog.showConfirm('删除文件夹', text.join(''), function(ok) {
-            if (ok) {
-                currentDir.deleteDirectory([ dir ], true, function(workingDir, resultList) {
-                    g.dialog.launchToast(Toast.Success, '已删除文件夹“' + dir.getName() + "”");
-
-                    that.refreshTable(true);
-                }, function(error) {
-                    g.dialog.launchToast(Toast.Error, '删除文件夹失败: ' + error.code);
-                });
+        if (selectedRecycleBin) {
+            var trashDir = g.cube().fs.queryTrash(dirId);
+            if (null == trashDir) {
+                alert('查找废弃目录出错');
+                return;
             }
-        }, '删除文件夹');
+
+            var text = ['您确定要删除文件夹 ', '“<span class="text-danger">', trashDir.getName(), '</span>” 及该文件夹内的',
+                        '<span class="text-danger">全部文件</span>',
+                        '吗？<p class="text-danger">该操作不可撤销！</p>'];
+            g.dialog.showConfirm('删除文件夹', text.join(''), function(ok) {
+                if (ok) {
+                    g.dialog.showLoading('删除文件夹');
+
+                    // 删除废弃文件
+                    g.cube().fs.eraseTrash([ dirId ], function(root, list) {
+                        g.dialog.launchToast(Toast.Success, '已删除文件夹“' + trashDir.getName() + "”");
+
+                        that.refreshTable(true, function() {
+                            g.dialog.hideLoading();
+                        });
+                    }, function(error) {
+                        g.dialog.hideLoading();
+                        g.dialog.launchToast(Toast.Error, '删除文件夹失败: ' + error.code);
+                    });
+                }
+            }, '删除文件夹');
+        }
+        else {
+            var dir = g.cube().fs.queryDirectory(dirId);
+            if (null == dir) {
+                alert('查找目录出错');
+                return;
+            }
+    
+            var text = ['您确定要删除文件夹 ', '“<span class="text-danger">', dir.getName(), '</span>” 及该文件夹内的',
+                        '<span class="text-danger">全部文件</span>',
+                        '吗？'];
+            g.dialog.showConfirm('删除文件夹', text.join(''), function(ok) {
+                if (ok) {
+                    currentDir.deleteDirectory([ dir ], true, function(workingDir, resultList) {
+                        g.dialog.launchToast(Toast.Success, '已删除文件夹“' + dir.getName() + "”");
+    
+                        that.refreshTable(true);
+                    }, function(error) {
+                        g.dialog.launchToast(Toast.Error, '删除文件夹失败: ' + error.code);
+                    });
+                }
+            }, '删除文件夹');
+        }
     }
 
     /**
@@ -10404,17 +10457,50 @@
         }
     }
 
+    /**
+     * 提示删除文件。
+     * @param {string} fileName 
+     * @param {string} fileCode 
+     */
     FilePanel.prototype.promptDeleteFile = function(fileName, fileCode) {
-        var text = ['您确定要删除文件 ', '“<span class="text-danger">', fileName, '</span>” 吗？'];
-        g.dialog.showConfirm('删除文件', text.join(''), function(ok) {
-            if (ok) {
-                currentDir.deleteFiles([ fileCode ], function(workingDir, resultList) {
-                    that.refreshTable(true);
-                }, function(error) {
-                    g.dialog.launchToast(Toast.Error, '删除文件失败: ' + error.code);
-                });
+        if (selectedRecycleBin) {
+            var trashFile = g.cube().fs.queryTrashFile(fileCode);
+            if (null == trashFile) {
+                return;
             }
-        }, '删除');
+
+            var text = ['您确定要删除文件 ', '“<span class="text-danger">', fileName, '</span>” 吗？',
+                        '<p class="text-danger">该操作不可撤销！</p>'];
+            g.dialog.showConfirm('删除文件', text.join(''), function(ok) {
+                if (ok) {
+                    g.dialog.showLoading('删除文件');
+
+                    g.cube().fs.eraseTrash([ trashFile.getId() ], function(root, list) {
+                        that.refreshTable(false);
+                        g.dialog.hideLoading();
+                    }, function(error) {
+                        g.dialog.hideLoading();
+                        g.dialog.launchToast(Toast.Error, '删除文件失败: ' + error.code);
+                    });
+                }
+            }, '删除');
+        }
+        else {
+            var text = ['您确定要删除文件 ', '“<span class="text-danger">', fileName, '</span>” 吗？'];
+            g.dialog.showConfirm('删除文件', text.join(''), function(ok) {
+                if (ok) {
+                    g.dialog.showLoading('删除文件');
+
+                    currentDir.deleteFiles([ fileCode ], function(workingDir, resultList) {
+                        that.refreshTable(true);
+                        g.dialog.hideLoading();
+                    }, function(error) {
+                        g.dialog.hideLoading();
+                        g.dialog.launchToast(Toast.Error, '删除文件失败: ' + error.code);
+                    });
+                }
+            }, '删除');
+        }
     }
 
     /**
